+++ NEWS.md
@@ -4,7 +4,7 @@ output: html_document
 ---
 
 
-# ECCTools 0.1.3
+# ECCTools 0.1.3 (14/07/2022)
 
 * exports_to_balancing function and corresponding tests added
 * losses_to_balancing function and corresponding tests added
@@ -12,6 +12,9 @@ output: html_document
 * specify_other_elec_heat_production function and corresponding tests added
 * specify_elect_heat_markets function and corresponding tests modified so that it works with the new specify_other_elec_heat_production function too
 * New specify_losses_as_industry function to treat losses as a transformation industry
+* ECC figure for country A updated
+* All tests passing (332)
+* Test coverage at 100%.
 
 
 # ECCTools 0.1.2 (07/01/2022)